Type fixes (16 -> 32 bit)

+++ include/AutonomousSystem.h
@@ -67,7 +67,7 @@ class AutonomousSystem : public GenericHashEntry,
 
   void set_hash_entry_state_idle();
 
-  inline u_int16_t getNumHosts() { return getUses(); }
+  inline u_int32_t getNumHosts() { return getUses(); }
   inline u_int32_t key() { return (asn); }
   inline u_int32_t get_asn() { return (asn); }
   inline char* get_asname() { return (asname ? asname : (char*)""); }
